Add handler, service and repository codes for us-0301

+++ basketapp/repository/migrations/20241103142601-create_basket_item_table.sql
@@ -1,7 +1,7 @@
 
 -- +migrate Up
 CREATE TABLE basket_items (
-    id BIGINT PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     basket_id BIGINT REFERENCES baskets(id) ON DELETE CASCADE,  
     food_id BIGINT NOT NULL,  
     food_name TEXT NOT NULL,
